+++ src/control/server/ctl_storage.go
@@ -28,9 +28,39 @@ import (
 
 	"github.com/pkg/errors"
 
-	"github.com/daos-stack/daos/src/control/common"
 	types "github.com/daos-stack/daos/src/control/common/storage"
+	"github.com/daos-stack/daos/src/control/drpc"
+	"github.com/daos-stack/daos/src/control/logging"
 )
+
+// StorageControlService encapsulates the storage part of the control service
+type StorageControlService struct {
+	log  logging.Logger
+	nvme *nvmeStorage
+	scm  *scmStorage
+	drpc drpc.DomainSocketClient
+}
+
+// NewStorageControlService returns an initialized *StorageControlService
+func NewStorageControlService(log logging.Logger, cfg *Configuration) (*StorageControlService, error) {
+	scriptPath, err := cfg.ext.getAbsInstallPath(spdkSetupPath)
+	if err != nil {
+		return nil, err
+	}
+
+	spdkScript := &spdkSetup{
+		log:         log,
+		scriptPath:  scriptPath,
+		nrHugePages: cfg.NrHugepages,
+	}
+
+	return &StorageControlService{
+		log:  log,
+		nvme: newNvmeStorage(log, cfg.NvmeShmID, spdkScript, cfg.ext),
+		scm:  newScmStorage(log, cfg.ext),
+		drpc: getDrpcClientConnection(cfg.SocketDir),
+	}, nil
+}
 
 type PrepareNvmeRequest struct {
 	HugePageCount int
